fix: seed lazily materialised zones with the cluster nameserver set
Writer.applyZone materialised a missing zone with a single ns1 pointing at
the writing node, while BootstrapBaseZone seeds ns1..nsN across every
northstar node. Both are create-if-absent and EnsureBaseZone never
overwrites, so on multi-node a RunInstances landing before the bootstrap
(which can wait ~60s on predastore) pinned compute.internal to one node's
degenerate NS topology for the life of the zone object.

Share one derivation rather than have the two agree by convention. As
services/northstar already imports handlers/dns, the derivation moves down
to handlers/dns/nameservers.go, joining the cluster-config-derived DNS
helpers vpcd already reads from there; bootstrap.go calls into it. This
also drops writer.go's localNameserverIP, a duplicate of nameserverIP.
