+++ spinifex/services/northstar/bootstrap.go
@@ -3,13 +3,13 @@ package northstar
 import (
 	"fmt"
 	"log/slog"
-	"net"
 	"sort"
 	"strings"
 	"time"
 
 	nsconfig "github.com/mulgadc/northstar/pkg/config"
 	"github.com/mulgadc/spinifex/spinifex/config"
+	handlers_dns "github.com/mulgadc/spinifex/spinifex/handlers/dns"
 )
 
 // baseZoneTXT is the marker TXT record seeded at the apex of the base zone.
@@ -76,7 +76,7 @@ func BootstrapBaseZone(configPath string, cluster *config.ClusterConfig) error {
 		SecretKey: sysCreds.SecretKey,
 	}
 
-	nameservers := buildNameserverSeeds(cluster)
+	nameservers := handlers_dns.NameserverSeeds(cluster)
 	slog.Info("northstar bootstrap: ensuring base zone",
 		"domain", domain, "nameservers", len(nameservers), "multi_node", len(nameservers) > 1)
 
@@ -141,71 +141,4 @@ func nodeNames(cluster *config.ClusterConfig) []string {
 	}
 	sort.Strings(names)
 	return names
-}
-
-// ResolverNameserverIPs returns the WAN IPs of cluster nodes running northstar,
-// in the same deterministic order as the seeded nameservers. vpcd's per-tap DNS
-// shim uses these as forward targets (northstar's :5300 listener), so internal
-// names resolve authoritatively and external names via upstream forwarders.
-// Loopback is skipped: a dev/misconfig node with no reachable IP yields an empty
-// list, letting the caller fall back to the upstream pool DNS.
-func ResolverNameserverIPs(cluster *config.ClusterConfig) []string {
-	names := configuredNorthstarNodeNames(cluster)
-	ips := make([]string, 0, len(names))
-	for _, name := range names {
-		ip := nameserverIP(cluster.Nodes[name])
-		if ip != "" && ip != "127.0.0.1" {
-			ips = append(ips, ip)
-		}
-	}
-	return ips
-}
-
-// configuredNorthstarNodeNames returns the deterministically ordered nodes that
-// explicitly advertise a northstar configuration.
-func configuredNorthstarNodeNames(cluster *config.ClusterConfig) []string {
-	var names []string
-	for name, node := range cluster.Nodes {
-		if node.Northstar.ConfigPath != "" {
-			names = append(names, name)
-		}
-	}
-	sort.Strings(names)
-	return names
-}
-
-// buildNameserverSeeds derives one nameserver (nsN → node IP) per cluster node
-// that runs northstar, ordered deterministically. Falls back to the local node
-// when no node advertises a northstar config (single-node / dev).
-func buildNameserverSeeds(cluster *config.ClusterConfig) []nsconfig.NameserverSeed {
-	names := configuredNorthstarNodeNames(cluster)
-	if len(names) == 0 {
-		names = []string{cluster.Node}
-	}
-
-	seeds := make([]nsconfig.NameserverSeed, 0, len(names))
-	for i, name := range names {
-		seeds = append(seeds, nsconfig.NameserverSeed{
-			Host: fmt.Sprintf("ns%d", i+1),
-			IP:   nameserverIP(cluster.Nodes[name]),
-		})
-	}
-	return seeds
-}
-
-// nameserverIP returns a reachable DNS address for a node: its advertised IP,
-// else its host, with any port stripped. A missing or wildcard address falls
-// back to loopback for single-node/dev.
-func nameserverIP(node config.Config) string {
-	ip := strings.TrimSpace(node.AdvertiseIP)
-	if ip == "" {
-		ip = strings.TrimSpace(node.Host)
-	}
-	if host, _, err := net.SplitHostPort(ip); err == nil {
-		ip = host
-	}
-	if ip == "" || ip == "0.0.0.0" {
-		ip = "127.0.0.1"
-	}
-	return ip
 }
