+++ spinifex/services/northstar/bootstrap_test.go
@@ -18,74 +18,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestNameserverIP(t *testing.T) {
-	assert.Equal(t, "10.11.12.1", nameserverIP(config.Config{AdvertiseIP: "10.11.12.1"}))
-	assert.Equal(t, "10.11.12.2", nameserverIP(config.Config{Host: "10.11.12.2:8443"}))
-	assert.Equal(t, "10.0.0.9", nameserverIP(config.Config{AdvertiseIP: "10.0.0.9", Host: "10.0.0.1"}))
-	assert.Equal(t, "127.0.0.1", nameserverIP(config.Config{Host: "0.0.0.0"}))
-	assert.Equal(t, "127.0.0.1", nameserverIP(config.Config{}))
-}
-
-func TestBuildNameserverSeeds(t *testing.T) {
-	// Single node with no northstar config path → fall back to the local node.
-	single := &config.ClusterConfig{
-		Node:  "node1",
-		Nodes: map[string]config.Config{"node1": {Host: "0.0.0.0"}},
-	}
-	seeds := buildNameserverSeeds(single)
-	require.Len(t, seeds, 1)
-	assert.Equal(t, "ns1", seeds[0].Host)
-	assert.Equal(t, "127.0.0.1", seeds[0].IP)
-
-	// Multi-node: one nameserver per node that advertises a northstar config,
-	// ordered deterministically.
-	multi := &config.ClusterConfig{
-		Node: "node1",
-		Nodes: map[string]config.Config{
-			"node2": {AdvertiseIP: "10.0.0.2", Northstar: config.NorthstarConfig{ConfigPath: "/etc/n.toml"}},
-			"node1": {AdvertiseIP: "10.0.0.1", Northstar: config.NorthstarConfig{ConfigPath: "/etc/n.toml"}},
-			"node3": {AdvertiseIP: "10.0.0.3"}, // no northstar → excluded
-		},
-	}
-	seeds = buildNameserverSeeds(multi)
-	require.Len(t, seeds, 2)
-	assert.Equal(t, "ns1", seeds[0].Host)
-	assert.Equal(t, "10.0.0.1", seeds[0].IP)
-	assert.Equal(t, "ns2", seeds[1].Host)
-	assert.Equal(t, "10.0.0.2", seeds[1].IP)
-}
-
-func TestResolverNameserverIPs(t *testing.T) {
-	// Multi-node: WAN IPs of the northstar nodes, ordered, non-northstar excluded.
-	multi := &config.ClusterConfig{
-		Node: "node1",
-		Nodes: map[string]config.Config{
-			"node2": {AdvertiseIP: "192.168.1.32", Northstar: config.NorthstarConfig{ConfigPath: "/etc/n.toml"}},
-			"node1": {AdvertiseIP: "192.168.1.31", Northstar: config.NorthstarConfig{ConfigPath: "/etc/n.toml"}},
-			"node3": {AdvertiseIP: "192.168.1.33"}, // no northstar → excluded
-		},
-	}
-	assert.Equal(t, []string{"192.168.1.31", "192.168.1.32"}, ResolverNameserverIPs(multi))
-
-	// Resolver discovery must not reuse the bootstrap fallback: a reachable node
-	// without northstar configured is not a valid DNS backend.
-	disabled := &config.ClusterConfig{
-		Node: "node1",
-		Nodes: map[string]config.Config{
-			"node1": {AdvertiseIP: "192.168.1.31"},
-			"node2": {AdvertiseIP: "192.168.1.32"},
-		},
-	}
-	assert.Empty(t, ResolverNameserverIPs(disabled))
-
-	// Loopback-only nodes are also excluded so the caller falls back to the
-	// upstream pool DNS.
-	dev := &config.ClusterConfig{Node: "node1", Nodes: map[string]config.Config{
-		"node1": {Host: "0.0.0.0", Northstar: config.NorthstarConfig{ConfigPath: "/etc/n.toml"}},
-	}}
-	assert.Empty(t, ResolverNameserverIPs(dev))
-}
-
 // fakeS3 is a minimal mutable path-style S3 endpoint (HEAD/PUT/GET) for the
 // bootstrap happy-path test.
 func fakeS3(t *testing.T, bucket string) (endpoint string, objects map[string]string) {
